+++ src/type/npointer.cpp
@@ -38,7 +38,8 @@ nobject npointer::get_target(const nobject& pointer)
 
 npointer::npointer(uint8_t depth, bool pointing_is_const, const ntype* pointing_type,
                    std::string_view name, operations* ops)
-    : ntype(ntype::etype::epointer, sizeof(void*), alignof(void*), ops, name),
+    : ntype(ntype::etype::epointer, static_cast<uint32_t>(sizeof(void*)),
+            static_cast<uint32_t>(alignof(void*)), ops, name),
       _depth(depth), _pointing_is_const(pointing_is_const), _pointing_type(pointing_type)
 {
 }
